# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateSubscriptionRequest:
    description: CreateSubscriptionRequest
    properties:
      end_date:
        example: 12-2025
        type: string
      price:
        example: 400
        minimum: 0
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 07-2025
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  dto.SubscriptionOutput:
    description: SubscriptionOutput
    properties:
      created_at:
        example: "2025-04-05T10:00:00Z"
        type: string
      end_date:
        example: 12-2025
        type: string
      id:
        example: a1b2c3d4-e5f6-7890-g1h2-i3j4k5l6m7n8
        type: string
      price:
        example: 400
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 07-2025
        type: string
      updated_at:
        example: "2025-04-05T10:00:00Z"
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  dto.SubscriptionsOutput:
    description: SubscriptionsOutput
    properties:
      has_next_page:
        example: true
        type: boolean
      has_prev_page:
        example: false
        type: boolean
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      subscriptions:
        items:
          $ref: '#/definitions/dto.SubscriptionOutput'
        type: array
      total:
        example: 42
        type: integer
    type: object
  dto.UpdateSubscriptionRequest:
    description: UpdateSubscriptionRequest
    properties:
      end_date:
        example: 06-2026
        type: string
      price:
        example: 599
        minimum: 0
        type: integer
      service_name:
        example: Spotify Premium
        type: string
      start_date:
        example: 01-2026
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for managing user subscriptions
  title: Subscription API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: Get paginated list of subscriptions with optional filters
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        type: string
      - description: Service name
        in: query
        name: service_name
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        maximum: 100
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionsOutput'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get list of subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create a subscription record for a user
      parameters:
      - description: Subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionOutput'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Delete a subscription by ID
      parameters:
      - description: Subscription ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a subscription
      tags:
      - subscriptions
    get:
      description: Retrieve a single subscription by its UUID
      parameters:
      - description: Subscription ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionOutput'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get subscription by ID
      tags:
      - subscriptions
    patch:
      consumes:
      - application/json
      description: Partially update a subscription by ID
      parameters:
      - description: Subscription ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionOutput'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a subscription
      tags:
      - subscriptions
  /subscriptions/cost:
    get:
      description: Calculate total cost of subscriptions for a given period and filters
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        type: string
      - description: Service name
        in: query
        name: service_name
        type: string
      - description: Start date in MM-YYYY format
        in: query
        name: start_date
        type: string
      - description: End date in MM-YYYY format
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Calculate total subscription cost
      tags:
      - subscriptions
schemes:
- http
swagger: "2.0"